atoi

+++ pointers_arrays_strings/100-atoi.c
@@ -7,7 +7,7 @@
  */
 int _atoi(char *s)
 {
-int sig = 1, num = 0, resp;
+double sig = 1, num = 0, resp;
 while (*s && (*s < '0' || *s > '9'))
 {
 if (*s == '-')
